+++ src/domain/controllers/index.ts
@@ -1,8 +1,9 @@
 import ItemsController from './itemsController'
 import RolesController from './rolesController'
 import UsersController from './usersController'
+import GroupItemsController from './groupItemsController'
 import Controller from './controller'
 
 export {
-  ItemsController, UsersController, RolesController, Controller
+  ItemsController, UsersController, RolesController, GroupItemsController, Controller
 }